fix(#95): use i64 for JWT timestamps instead of usize
Utc::now().timestamp() as usize and usize exp/iat claims overflow after 2038
on 32-bit targets and can be UB for negative timestamps. Switch claims and
derived values to i64 (chrono's native type), removing all timestamp casts.
revocation_ttl_secs now clamps with .max(0) since i64 saturating_sub
saturates at i64::MIN, not 0.

+++ crates/auth/src/adapters/google_auth_provider.rs
@@ -217,8 +217,10 @@ pub struct GoogleClaims {
     pub name: Option<String>,
     pub iss: String,
     pub aud: String,
-    pub exp: usize,
-    pub iat: usize,
+    /// Unix seconds; `i64` for the same platform/overflow reasons as the app's
+    /// own JWT claims (see #95).
+    pub exp: i64,
+    pub iat: i64,
 }
 
 #[cfg(test)]
